Crate lmdb_sys

Structs

MDB_envinfo

MDB_stat

MDB_val

Enums

MDB_cursor

MDB_env

MDB_txn

Constants

MDB_APPEND
Data is being appended, don't split full pages.

MDB_APPENDDUP
Duplicate data is being appended, don't split full pages.

MDB_BAD_DBI
The specified DBI was changed unexpectedly.

MDB_BAD_RSLOT
Invalid reuse of reader locktable slot.

MDB_BAD_TXN
Transaction cannot recover - it must be aborted.

MDB_BAD_VALSIZE
Unsupported size of key/DB name/data, or wrong DUPFIXED size.

MDB_CORRUPTED
Located page was wrong type.

MDB_CP_COMPACT
Compacting copy: Omit free space from copy, and renumber all pages sequentially.

MDB_CREATE
create DB if not already existing.

MDB_CURRENT
For mdb_cursor_put: overwrite the current key/data pair.

MDB_CURSOR_FULL
Cursor stack too deep - internal error.

MDB_DBS_FULL
Environment maxdbs reached.

MDB_DUPFIXED
with MDB_DUPSORT, sorted dup items have fixed size.

MDB_DUPSORT
use sorted duplicates

MDB_FIRST

MDB_FIRST_DUP

MDB_FIXEDMAP
mmap at a fixed address (experimental)

MDB_GET_BOTH

MDB_GET_BOTH_RANGE

MDB_GET_CURRENT

MDB_GET_MULTIPLE

MDB_INCOMPATIBLE
MDB_INCOMPATIBLE: Operation and DB incompatible, or DB flags changed.

MDB_INTEGERDUP
with MDB_DUPSORT, dups are numeric in native byte order.

MDB_INTEGERKEY
numeric keys in native byte order. The keys must all be of the same size.

MDB_INVALID
File is not a valid LMDB file.

MDB_KEYEXIST
key/data pair already exists.

MDB_LAST

MDB_LAST_DUP

MDB_LAST_ERRCODE
The last defined error code.

MDB_MAPASYNC
use asynchronous msync when #MDB_WRITEMAP is used

MDB_MAP_FULL
Environment mapsize reached.

MDB_MAP_RESIZED
Database contents grew beyond environment mapsize.

MDB_MULTIPLE
Store multiple data items in one call. Only for #MDB_DUPFIXED.

MDB_NEXT

MDB_NEXT_DUP

MDB_NEXT_MULTIPLE

MDB_NEXT_NODUP

MDB_NODUPDATA
Only for MDB_DUPSORT.

MDB_NOLOCK
don't do any locking, caller must manage their own locks

MDB_NOMEMINIT
don't initialize malloc'd memory before writing to datafile

MDB_NOMETASYNC
don't fsync metapage after commit

MDB_NOOVERWRITE
For put: Don't write if the key already exists.

MDB_NORDAHEAD
don't do readahead (no effect on Windows)

MDB_NOSUBDIR
no environment directory

MDB_NOSYNC
don't fsync after commit

MDB_NOTFOUND
key/data pair not found (EOF).

MDB_NOTLS
tie reader locktable slots to #MDB_txn objects instead of to threads

MDB_PAGE_FULL
Page has not enough space - internal error.

MDB_PAGE_NOTFOUND
Requested page not found - this usually indicates corruption.

MDB_PANIC
Update of meta page failed or environment had fatal error.

MDB_PREV

MDB_PREV_DUP

MDB_PREV_NODUP

MDB_RDONLY
read only

MDB_READERS_FULL
Environment maxreaders reached.

MDB_RESERVE
For put: Just reserve space for data, don't copy it. Return a pointer to the reserved space.

MDB_REVERSEDUP
with #MDB_DUPSORT, use reverse string dups.

MDB_REVERSEKEY
use reverse string keys

MDB_SET

MDB_SET_KEY

MDB_SET_RANGE

MDB_SUCCESS
Successful result.

MDB_TLS_FULL
Too many TLS keys in use - Windows only.

MDB_TXN_FULL
Txn has too many dirty pages.

MDB_VERSION_MISMATCH
Environment version mismatch.

MDB_WRITEMAP
use writable mmap

Functions

mdb_cmp

mdb_cursor_close

mdb_cursor_count

mdb_cursor_dbi

mdb_cursor_del

mdb_cursor_get

mdb_cursor_open

mdb_cursor_put

mdb_cursor_renew

mdb_cursor_txn

mdb_dbi_close

mdb_dbi_flags

mdb_dbi_open

mdb_dcmp

mdb_del

mdb_drop

mdb_env_close

mdb_env_copy

mdb_env_copy2

mdb_env_copyfd

mdb_env_copyfd2

mdb_env_create

mdb_env_get_fd

mdb_env_get_flags

mdb_env_get_maxkeysize

mdb_env_get_maxreaders

mdb_env_get_path

mdb_env_get_userctx

mdb_env_info

mdb_env_open

mdb_env_set_assert

mdb_env_set_flags

mdb_env_set_mapsize

mdb_env_set_maxdbs

mdb_env_set_maxreaders

mdb_env_set_userctx

mdb_env_stat

mdb_env_sync

mdb_get

mdb_put

mdb_reader_check

mdb_reader_list

mdb_set_compare

mdb_set_dupsort

mdb_set_relctx

mdb_set_relfunc

mdb_stat

mdb_strerror

mdb_txn_abort

mdb_txn_begin

mdb_txn_commit

mdb_txn_env

mdb_txn_id

mdb_txn_renew

mdb_txn_reset

mdb_version

Type Definitions

MDB_assert_func

MDB_cmp_func

MDB_cursor_op

MDB_dbi

MDB_msg_func

MDB_rel_func

mode_t